refactor: refactor ChatClientDelegate injection and add default empty implementation
