+++ spring-ai-alibaba-studio/src/main/java/com/alibaba/cloud/ai/controller/ChatClientAPIController.java
@@ -17,8 +17,10 @@ package com.alibaba.cloud.ai.controller;
 
 import com.alibaba.cloud.ai.api.ChatClientAPI;
 import com.alibaba.cloud.ai.service.ChatClientDelegate;
+import com.alibaba.cloud.ai.service.impl.EmptyChatClientDelegate;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -30,9 +32,10 @@ public class ChatClientAPIController implements ChatClientAPI {
 
 	private final ChatClientDelegate delegate;
 
-	public ChatClientAPIController(@Autowired(required = false) ChatClientDelegate delegate) {
-		this.delegate = Optional.ofNullable(delegate).orElse(new ChatClientDelegate() {
-		});
+	public ChatClientAPIController(
+			@Autowired(required = false) ChatClientDelegate delegate,
+			@Autowired @Qualifier("emptyChatClientDelegate") EmptyChatClientDelegate emptyDelegate) {
+		this.delegate = Optional.ofNullable(delegate).orElse(emptyDelegate);
 	}
 
 	@Override
